functions to build models

functions to build models

Usage

mod_funs(
  include = c("constIntOnly", "tvIntOnly", "tvSlope", "constLM", "tvCRzeroInt",
    "constCRzeroInt", "tvInt"),
  penDLM_formula = NULL,
  penDLM_regu = NULL,
  penDLM_gamma_shape = NULL,
  penDLM_gamma_scale = NULL
)

Arguments

include

vector of names of models to include. Default is all options

penDLM_formula

an optional r formula object to pass to the penalized DLM model. Default is NULL, which would use the defaulte value formula("y~x") from the pen_dlm function.

penDLM_regu

scaler for the addition of the log of standard deviations to the log-likelihood as a regularizer. See penealized_dlm help for more details

penDLM_gamma_shape

shape parameter for the gamma prior on the exponential distribution rate parameter. See penealized_dlm help for more details

penDLM_gamma_scale

scale parameter for the gamma prior on the exponential distribution rate parameter.See penealized_dlm help for more details

Value

a list of functions that return (unoptimized) dlm model objects

Details

Model options are:

constLM – Sibling regression with constant slope and intercept.

tvInt – Sibling regression with time-varying intercept.

tvSlope – Sibling regression with time-varying slope.

tvIntSlope – Sibling regression with time-varying slope and intercept. This model is not inlcuded by default because it generally converges to one of the simpler model. In other words, it is exceedingly rare for the data to support this complex of a model and this including it in an ensemble leads to double weighting one of the simpler models.

tvCRzeroInt – Time varying "cohort ratio" model. Time varying slope, Intercept=0.

constCRzeroInt – Constant "cohort ratio" model. Constant slope, Intercept=0.

tvIntOnly – Time-varying Intercept-only model. Random walk on return, no sibling predictor.

constIntOnly – Constant Intercept-only model. Long-term average, no sibling predictor.

PenDlm – A dynamic linear model the like "tvIntSlope" but with penalized-complexity priors on the year-to-year variation and the mean of the covariate values. This model also can accommodate additional predictors (e.g., environmental covariates). This model is not included by defaults