penalized DLM in RTMB

function to fit a penalized dynamic linear regression model.

The penalized complexity model puts penalties on the mean of the coefficients \(\beta_t\) for each year \(t\) and the standard deviation of the steps in the random walk. So if the coefficients are:

$$\beta_t = \beta_{t-1} + \omega_t \\ \omega_t \sim \mathcal{N}(0,\sigma)$$

this model puts exponential-gamma penalties on both \(\bar{\beta}\) and \(\sigma\).

$$\bar{\beta},~\sigma \sim \text{exp}(\lambda) \\ \lambda \sim \text{Gamma}(\text{Shape}=10,~\text{Scale}=1)$$

Additionally, \(0.05*\text{log}(\bar{\beta}, \sigma)\) for all \(\bar{\beta}\)s and \(\sigma\)s is added to the log-likelihood to keep their values from shrinking so small as to cause numerical problems during optimization.

Usage

pen_dlm(
  dat,
  form = stats::formula("y~x"),
  regu = c(0.05, 0.05),
  gamma_shape = 10,
  gamma_scale = 1
)

Arguments

dat

dataframe with response and predictors

form

formula specifying linear predictor

regu

to avoid the standard deviations shrinking too small and running into numerical issues, the log of the standard deviation multiplied be these values are added to the log-likelihood (subtracted from the negative log-likelihood). Defaults are .05 and .05 for the penalties on the mean of the coefficients and the year-to-year variability.

gamma_shape

shape paramter for the gamma prior on the expential distributions rate paramters.

gamma_scale

scale paramter for the gamma prior on the expential distributions rate paramters.

Value

a list with two components: the fitted TMB model object, which has the NLL and the linear predictors in the report(), as well as the outfut from the call to TMBhelper::fit_tmb(), which is used to optimize the model.